Fix panic when request body is nil
The stdlib transport checks certain attributes of the reqest body
by probing some of the methods and checking the responses. One of
these methods is .Read() which, after being wrapped by the middleware,
calls down to a nil pointer if the original request body was nil.

This happens because the body wrapper was being applied in the nil
case. Now, if the original request body is nil we don't modify it
and calculate the bytes read as zero.